Damage System not working

+++ framework/src/main/java/no/ntnu/idi/apollo69framework/network_messages/data_transfer_objects/AsteroidDto.java
@@ -3,12 +3,12 @@ package no.ntnu.idi.apollo69framework.network_messages.data_transfer_objects;
 public class AsteroidDto {
     public PositionDto positionDto;
     public VelocityDto velocityDto;
-    public int hp;
+    public float hp;
 
     public AsteroidDto() {
     }
 
-    public AsteroidDto(PositionDto positionDto, int hp) {
+    public AsteroidDto(PositionDto positionDto, float hp) {
         this.positionDto = positionDto;
         this.hp = hp;
     }
